[feat #12] 스크랩 entity 구조 변경
