+++ src/main/java/com/umc/ttt/domain/scrap/entity/BookScrap.java
@@ -1,0 +1,27 @@
+package com.umc.ttt.domain.scrap.entity;
+
+import com.umc.ttt.domain.book.entity.Book;
+import com.umc.ttt.global.common.BaseEntity;
+import jakarta.persistence.*;
+import lombok.*;
+
+@Builder
+@Getter
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+@AllArgsConstructor(access = AccessLevel.PRIVATE)
+@Entity
+public class BookScrap extends BaseEntity {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    @Column(name = "book_scrap_id", nullable = false)
+    private Long id;
+
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "book_id")
+    private Book book;
+
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "scrap_folder_id")
+    private ScrapFolder scrapFolder;
+}
